Rename common auth middleware
This change renames the middleware to be more descriptive of what
it is doing.

A change will be coming to determine whether a particular
route can be accessed or not so this will avoid any name confusion.
