+++ server.js
@@ -26,7 +26,7 @@ const redisStore = require('./config/redis-store')
 const currentLocation = require('./common/middleware/current-location')
 const errorHandlers = require('./common/middleware/errors')
 const checkSession = require('./common/middleware/check-session')
-const ensureAuthenticated = require('./common/middleware/authentication')
+const ensureAuthenticated = require('./common/middleware/ensure-authenticated')
 const locals = require('./common/middleware/locals')
 const router = require('./app/router')
 const healthcheckApp = require('./app/healthcheck')
